Complete home and projects layout
